requirements.txt
# Core GenAI modules
transformers==4.35.2
diffusers==0.24.0
sentence-transformers==2.2.2
accelerate==0.23.0
huggingface-hub==0.20.2

# Vision (BLIP-2 captioning)
torch>=1.13.0  # or your installed version
Pillow>=9.0.0

# PDF Processing
PyMuPDF==1.23.7  # known stable with modern Python

# Audio (fallback TTS using pyttsx3)
pyttsx3==2.90

# Vector Search
faiss-cpu==1.7.4

# App Framework
streamlit==1.33.0

# Utils
numpy>=1.21
requests
tqdm
regex
pyyaml
uuid
